+++ libs/audio/cd/cd_sgi.c
@@ -268,12 +268,12 @@ pCD_f (void)
 	}
 
 	if (strequal (command, "play")) {
-		CDAudio_Play ((int) atoi (Cmd_Argv (2)), false);
+		CDAudio_Play (atoi (Cmd_Argv (2)), false);
 		return;
 	}
 
 	if (strequal (command, "loop")) {
-		CDAudio_Play ((int) atoi (Cmd_Argv (2)), true);
+		CDAudio_Play (atoi (Cmd_Argv (2)), true);
 		return;
 	}
